+++ quilt-tester/tests/10-network/network.js
@@ -4,18 +4,18 @@ let infrastructure = require('../../config/infrastructure.js');
 let deployment = quilt.createDeployment();
 deployment.deploy(infrastructure);
 
-let c = new quilt.Container('alpine', {
-    command: ['tail', '-f', '/dev/null'],
-});
+const image = 'alpine';
+const command = ['tail', '-f', '/dev/null'];
 
-let red = new quilt.Service('red', c.replicate(5));
-setHostnames(red.containers, 'red');
-
-let blue = new quilt.Service('blue', c.replicate(5));
-setHostnames(blue.containers, 'blue');
-
-let yellow = new quilt.Service('yellow', c.replicate(5));
-setHostnames(blue.containers, 'blue');
+let red = new quilt.Service('red-lb',
+  new quilt.Container('red', image, {command}).replicate(5)
+);
+let blue = new quilt.Service('blue-lb',
+  new quilt.Container('blue', image, {command}).replicate(5)
+);
+let yellow = new quilt.Service('yellow-lb',
+  new quilt.Container('yellow', image, {command}).replicate(5)
+);
 
 blue.allowFrom(red, 80);
 red.allowFrom(blue, 80);
@@ -25,16 +25,3 @@ yellow.allowFrom(blue, 80);
 deployment.deploy(red);
 deployment.deploy(blue);
 deployment.deploy(yellow);
-
-/**
- * setHostnames sets the hostnames for `containers` to be a unique hostname
- * prefixed by `hostname`.
- *
- * @param {quilt.Container[]} containers
- * @param {string} hostname
- */
-function setHostnames(containers, hostname) {
-    containers.forEach((c) => {
-        c.setHostname(hostname);
-    });
-}