+++ quilt-tester/tests/connection-credentials/connection-credentials.js
@@ -4,5 +4,7 @@ let infrastructure = require('../../config/infrastructure.js');
 let deployment = quilt.createDeployment();
 deployment.deploy(infrastructure);
 
-let red = new quilt.Service('red', [new quilt.Container('google/pause')]);
+let red = new quilt.Service('red', [
+  new quilt.Container('red', 'google/pause'),
+]);
 deployment.deploy(red);