+++ quilt-tester/tests/load-balancer/load-balancer.js
@@ -6,14 +6,15 @@ deployment.deploy(infrastructure);
 
 let containers = [];
 for (let i = 0; i < 4; i++) {
-  containers.push(new quilt.Container('nginx:1.10').withFiles({
+  containers.push(new quilt.Container('web', 'nginx:1.10').withFiles({
     '/usr/share/nginx/html/index.html':
         'I am container number ' + i.toString() + '\n',
   }));
 }
 
 let fetcher = new quilt.Service('fetcher',
-    [new quilt.Container('alpine', {command: ['tail', '-f', '/dev/null']})]);
+    [new quilt.Container('fetcher', 'alpine', {
+        command: ['tail', '-f', '/dev/null']})]);
 let loadBalanced = new quilt.Service('loadBalanced', containers);
 loadBalanced.allowFrom(fetcher, 80);