+++ quilt-tester/tests/outbound-public/outbound-public.js
@@ -5,14 +5,16 @@ let deployment = quilt.createDeployment();
 deployment.deploy(infrastructure);
 
 let connected = new quilt.Service('connected',
-    new quilt.Container('alpine', {command: ['tail', '-f', '/dev/null']})
-        .replicate(infrastructure.nWorker*2)
+    new quilt.Container('outbound', 'alpine', {
+        command: ['tail', '-f', '/dev/null'],
+    }).replicate(infrastructure.nWorker*2)
 );
 quilt.publicInternet.allowFrom(connected, 80);
 
 let notConnected = new quilt.Service('not-connected',
-    new quilt.Container('alpine', {command: ['tail', '-f', '/dev/null']})
-        .replicate(infrastructure.nWorker*2)
+    new quilt.Container('ignoreme', 'alpine', {
+        command: ['tail', '-f', '/dev/null'],
+    }).replicate(infrastructure.nWorker*2)
 );
 
 deployment.deploy([connected, notConnected]);